Refactor LocalPackageDto and update serialization to use MemoryPack
