--- Shelly-CLI/Commands/Standard/LocalPackageDto.cs
@@ -1,7 +1,0 @@
-namespace Shelly_CLI.Commands.Standard;
-
-// TODO: Move to PackageManager #771
-public record LocalPackageDto(
-    string Name,
-    long Size
-);
